@@ -5,8 +5,8 @@ import net.n2oapp.platform.selection.api.Selection;
 import net.n2oapp.platform.selection.integration.model.EmployeeSelection;
 import org.springframework.data.domain.Sort;
 
-import javax.ws.rs.BeanParam;
-import javax.ws.rs.QueryParam;
+import jakarta.ws.rs.BeanParam;
+import jakarta.ws.rs.QueryParam;
 import java.util.List;
 
 public class EmployeeCriteria extends RestCriteria {
