@@ -3,8 +3,8 @@ package net.n2oapp.platform.selection.integration.rest;
 import net.n2oapp.platform.selection.integration.model.Employee;
 import org.springframework.data.domain.Page;
 
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
 
 @Path("/selective")
 @Consumes({MediaType.APPLICATION_JSON})
