@@ -1,8 +1,8 @@
 package net.n2oapp.platform.loader.server;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.jaxrs.json.JacksonJaxbJsonProvider;
-import com.fasterxml.jackson.jaxrs.json.JacksonJsonProvider;
+import com.fasterxml.jackson.jakarta.rs.json.JacksonXmlBindJsonProvider;
+import com.fasterxml.jackson.jakarta.rs.json.JacksonJsonProvider;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
@@ -12,7 +12,7 @@ import java.util.List;
 class TestApplication {
     @Bean
     JacksonJsonProvider jsonProvider(ObjectMapper objectMapper) {
-        return new JacksonJsonProvider(JacksonJaxbJsonProvider.DEFAULT_ANNOTATIONS);
+        return new JacksonJsonProvider(JacksonXmlBindJsonProvider.DEFAULT_ANNOTATIONS);
     }
 
     @Bean
